requirements.txt
fastapi
uvicorn[standard]
torch       
torchvision
huggingface_hub
pillow